new scripts change_updatedbn.sh will take an csv in first,last,did format, pull current config, diff the two files, output an import file for vm-move script, output a file to show same users, and remove current config and replace with the provided list
file_move-users.sh takes a csv input in first,last,did format and will
remove those lines from accounts.csv and readd them to the bottom of the
file, can comment out lines in the file to add to top of the file, so
can be used with flags for moving users to the top for recording at some point.
right now its manually commented out.

minor changes to other scripts to allow the usage inside these scripts.